discount/pyproject.toml
[tool.poetry]
name = "discount"
version = "0.1.0"
description = "GRPC discount microservice"
authors = ["Miha Sedej <sedej.miha@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.22"
grpcio-tools = "^1.22"

[tool.poetry.dev-dependencies]
black = { version = "*", allows-prereleases = true }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
